requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
supabase
psycopg2-binary
python-dotenv
elevenlabs
pydantic
langchain-postgres
pydantic-settings
unstructured
markdown
prompt_toolkit
